refactor: decompose MCP server tool registrars
