+++ src/mcp/types.ts
@@ -1,0 +1,7 @@
+export interface WorkgraphMcpServerOptions {
+  workspacePath: string;
+  defaultActor?: string;
+  readOnly?: boolean;
+  name?: string;
+  version?: string;
+}
